je créé une ligne de commande !

+++ back/src/main/java/im2agoracle/univgrenoblealpes/fr/gromed/repository/CommandeRepository.java
@@ -12,5 +12,5 @@ import im2agoracle.univgrenoblealpes.fr.gromed.entities.Commande;
 public interface CommandeRepository extends JpaRepository<Commande, Integer>{
     
     @Query(value="select c from commande c where c.idetablissement = ?1 and estpanier = 1", nativeQuery = true)
-    Optional<Commande> findPanierByIdEtab(int idEtab);
+    Optional<Commande> findPanierByIdEtab(long idEtab);
 }
